added rel_model and swagger in labelstudio ml backend, added migrations for clickhouse
