+++ dev_tools/label_studio/data/export/project-1-at-2025-05-17-19-33-fbe8476e-info.json
@@ -1,0 +1,1 @@
+{"project":{"title":"Patents NER","id":1,"created_at":"2025-04-18T13:43:31Z","created_by":"danil-nikonov-2017@mail.ru","task_number":1,"annotation_number":1},"platform":{"version":"1.18.0+0.ga5c4b5a"},"download":{"GET":{"exportType":["JSON"]},"time":"2025-05-17T19:33:51Z","result_filename":"\/label-studio\/data\/export\/project-1-at-2025-05-17-19-33-fbe8476e.json","md5":"fbe8476e864069a0a628b0ca81058695"}}
