+++ dev_tools/label_studio/data/export/project-1-at-2025-05-17-19-33-fbe8476e.json
@@ -1,0 +1,1 @@
+[{"id":1,"annotations":[{"id":2638,"completed_by":2,"result":[{"value":{"start":191,"end":201,"text":"Устройство","labels":["SYSTEM"]},"id":"30","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":211,"end":219,"text":"обечайку","labels":["COMPONENT"]},"id":"32","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":231,"end":244,"text":"верхней части","labels":["ATTRIBUTE"]},"id":"36","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":245,"end":265,"text":"перфорированный пояс","labels":["COMPONENT"]},"id":"38","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":267,"end":292,"text":"обруч определенной ширины","labels":["ATTRIBUTE"]},"id":"41","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":317,"end":325,"text":"обечайки","labels":["COMPONENT"]},"id":"47","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":327,"end":350,"text":"измеритель толщины слоя","labels":["COMPONENT"]},"id":"49","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":379,"end":404,"text":"привод перемещения обруча","labels":["COMPONENT"]},"id":"55","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":422,"end":430,"text":"обечайки","labels":["COMPONENT"]},"id":"60","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":531,"end":540,"text":"отверстий","labels":["COMPONENT"]},"id":"76","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":543,"end":551,"text":"обечайке","labels":["COMPONENT"]},"id":"78","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":637,"end":655,"text":"внутреннюю полость","labels":["COMPONENT"]},"id":"91","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":656,"end":664,"text":"обечайки","labels":["COMPONENT"]},"id":"93","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"value":{"start":700,"end":708,"text":"скиммера","labels":["COMPONENT"]},"id":"101","from_name":"label","to_name":"text","type":"labels","origin":"manual"},{"from_id":"32","to_id":"30","type":"relation","direction":"right","labels":["PART-OF"]},{"from_id":"36","to_id":"38","type":"relation","direction":"right","labels":["ATTRIBUTE-FOR"]},{"from_id":"38","to_id":"30","type":"relation","direction":"right","labels":["PART-OF"]},{"from_id":"47","to_id":"30","type":"relation","direction":"right","labels":["PART-OF"]},{"from_id":"49","to_id":"30","type":"relation","direction":"right","labels":["PART-OF"]},{"from_id":"55","to_id":"30","type":"relation","direction":"right","labels":["PART-OF"]}],"was_cancelled":false,"ground_truth":false,"created_at":"2025-05-17T19:24:49.281567Z","updated_at":"2025-05-17T19:28:27.767964Z","draft_created_at":null,"lead_time":188.41299999999998,"prediction":{},"result_count":15,"unique_id":"4ff88897-b992-4ef4-9d02-c9a62e75c948","import_id":null,"last_action":null,"bulk_created":false,"task":1,"project":1,"updated_by":2,"parent_prediction":null,"parent_annotation":null,"last_created_by":null}],"file_upload":"7a6cd1d4-tmp.txt","drafts":[],"predictions":[61],"data":{"text":"Полезная модель относится к устройствам для очистки водоемов от нефти и других подобных веществ, плавающих на водной поверхности, а именно к устройствам для изменения уровня порога скиммера. Устройство включает обечайку, имеющую в верхней части перфорированный пояс, обруч определенной ширины, смонтированный снаружи обечайки, измеритель толщины слоя собираемого нефтепродукта и привод перемещения обруча вдоль образующей обечайки. В процессе работы устройства обруч при своем перемещении либо приоткрывает, либо перекрывает часть отверстий в обечайке, соответственно, уменьшая или увеличивая толщину слоя нефтепродукта, поступающего во внутреннюю полость обечайки, т.е. регулирует величину «порога» скиммера."},"meta":{},"created_at":"2025-05-17T18:32:21.501351Z","updated_at":"2025-05-17T19:31:03.490882Z","inner_id":1,"total_annotations":1,"cancelled_annotations":0,"total_predictions":1,"comment_count":0,"unresolved_comment_count":0,"last_comment_updated_at":null,"project":1,"updated_by":2,"comment_authors":[]}]
