--- label_studio_ml_backend/RuPatents_ner/.dockerignore
@@ -1,19 +1,0 @@
-# Exclude everything
-**
-
-# Include Dockerfile and docker-compose for reference (optional, decide based on your use case)
-!Dockerfile
-!docker-compose.yml
-
-# Include Python application files
-!*.py
-
-# Include requirements files
-!requirements*.txt
-
-# Include script
-!*.sh
-
-# Exclude specific requirements if necessary
-# requirements-test.txt (Uncomment if you decide to exclude this)
-
