--- label_studio_ml_backend/RuPatents_ner/docker-compose.yml
@@ -1,29 +1,0 @@
-version: "3.8"
-
-services:
-  ml-backend:
-    container_name: ml-backend
-    image: humansignal/ml-backend:v0
-    build:
-      context: .
-      args:
-        TEST_ENV: ${TEST_ENV}
-        DEBUG: ${DEBUG}
-    environment:
-      - LOG_LEVEL=DEBUG
-      - ANY=PARAMETER
-      - WORKERS=1
-      - THREADS=8
-      - MODEL_DIR=/data/models
-    ports:
-      - "9091:9090"
-    volumes:
-      - "./data/server:/data"
-      - ./model.py:/app/model.py
-      - ./_wsgi.py:/app/_wsgi.py
-    networks:
-      - labelstudio
-
-networks:
-  labelstudio:
-    external: true
