--- label_studio_ml_backend/RuPatents_ner/requirements-base.txt
@@ -1,2 +1,0 @@
-gunicorn==22.0.0
-label-studio-ml @ git+https://github.com/HumanSignal/label-studio-ml-backend.git
