--- label_studio_ml_backend/RuPatents_ner/requirements-model.txt
@@ -1,3 +1,0 @@
-spacy==3.8.5
-spacy-transformers==1.3.8
-scikit-learn==1.6.1
