--- label_studio_ml_backend/RuPatents_ner/requirements-test.txt
@@ -1,2 +1,0 @@
-pytest
-pytest-cov
