--- label_studio_ml_backend/RuPatents_ner/requirements.txt
@@ -1,4 +1,0 @@
-spacy==3.8.5
-spacy-transformers==1.3.8
-scikit-learn==1.6.1
-ru_patents_ner @ https://huggingface.co/Delicalib/ru_patents_ner/resolve/main/ru_patents_ner-any-py3-none-any.whl
