+++ label_studio_ml_backend/RuPatents_rel/requirements-base.txt
@@ -1,0 +1,2 @@
+gunicorn==22.0.0
+label-studio-ml @ git+https://github.com/HumanSignal/label-studio-ml-backend.git
