+++ label_studio_ml_backend/RuPatents_rel/requirements-test.txt
@@ -1,0 +1,2 @@
+pytest
+pytest-cov
