+++ label_studio_ml_backend/RuPatents_rel/requirements.txt
@@ -1,0 +1,7 @@
+spacy==3.8.5
+spacy-transformers==1.3.8
+scikit-learn==1.6.1
+plotly==6.1.0
+clickhouse-connect==0.8.17
+Flask-RESTful==0.3.10
+flasgger==0.9.7.1
