--- label_studio_ml_backend/label_studio_ml.egg-info/SOURCES.txt
@@ -1,36 +1,0 @@
-LICENSE
-MANIFEST.in
-README.md
-setup.py
-label_studio_ml/__init__.py
-label_studio_ml/api.py
-label_studio_ml/cache.py
-label_studio_ml/exceptions.py
-label_studio_ml/ls_io.py
-label_studio_ml/model.py
-label_studio_ml/response.py
-label_studio_ml/server.py
-label_studio_ml/utils.py
-label_studio_ml/wsgi.py
-label_studio_ml.egg-info/PKG-INFO
-label_studio_ml.egg-info/SOURCES.txt
-label_studio_ml.egg-info/dependency_links.txt
-label_studio_ml.egg-info/entry_points.txt
-label_studio_ml.egg-info/requires.txt
-label_studio_ml.egg-info/top_level.txt
-label_studio_ml/default_configs/.dockerignore
-label_studio_ml/default_configs/Dockerfile
-label_studio_ml/default_configs/README.md
-label_studio_ml/default_configs/_wsgi.py.tmpl
-label_studio_ml/default_configs/docker-compose.yml
-label_studio_ml/default_configs/model.py
-label_studio_ml/default_configs/requirements-base.txt
-label_studio_ml/default_configs/requirements-test.txt
-label_studio_ml/default_configs/requirements.txt
-label_studio_ml/default_configs/test_api.py
-tests/__init__.py
-tests/test_api.py
-tests/test_basic.py
-tests/test_build_label_map.py
-tests/test_response.py
-tests/test_unit_model.py
