--- label_studio_ml_backend/label_studio_ml.egg-info/entry_points.txt
@@ -1,2 +1,0 @@
-[console_scripts]
-label-studio-ml = label_studio_ml.server:main
