--- label_studio_ml_backend/label_studio_ml.egg-info/requires.txt
@@ -1,6 +1,0 @@
-Flask>2.3
-colorama~=0.4
-requests~=2.31
-semver~=3.0.2
-pillow~=10.3
-label-studio-sdk@ git+https://github.com/HumanSignal/label-studio-sdk.git
