--- label_studio_ml_backend/label_studio_ml.egg-info/top_level.txt
@@ -1,2 +1,0 @@
-label_studio_ml
-tests
