+++ models_tools/REL model src/rel_component/configs/rel_trf.cfg
@@ -120,9 +120,10 @@ initial_rate = 5e-5
 rel_micro_p = 0.0
 rel_micro_r = 0.0
 rel_micro_f = 1.0
-f1_macro = 1.0
-f1_weighted = 1.0
+rel_macro_f = 1.0
+rel_weighted_f = 1.0
 f1_PART-OF = 1.0
 f1_LOCATED-AT = 1.0
 f1_CONNECTED-WITH = 1.0
 f1_IN-MANNER-OF = 1.0
+f1_ATTRIBUTE-FOR = 1.0
